requirements.txt
fastapi
uvicorn
agno
pydantic
bindufy